requirements.txt
pytorch
librosa
scikit-learn
pandas
torchaudio
